fix(agent): drop the live hook, and make the settled transcript honest
Review found that `on_agent_event` was not live, and it was right. The
transcript rides the outcome `AgentRunner::run` returns, so the hook could only
fire once the whole agent turn was over — in a burst, at the same moment
`on_step_finish` gets the same entries. For the long-running node it was
justified by, it published nothing until there was nothing left to publish.

Removing it rather than patching the docs. Genuine live delivery needs a sink on
the capability contract, and `AgentRunRequest` cannot carry one — it is
`Serialize` + `PartialEq` — so it wants a new trait method designed on purpose,
not a hook that already promises something it cannot do. Its only consumer does
not implement it, so nothing is lost today and the API stops lying.

Also from review:

- Per-item transcripts are keyed by item index and drained in that order.
  `map_items` restores outputs to input order; appending on completion let a
  `concurrency > 1` run interleave differently from the next for identical
  input.
- `TranscriptEntry::bounded` charges the truncation marker against
  `MAX_ENTRY_TEXT_BYTES` instead of adding it on top, so a clipped entry
  actually honours the documented cap.
- `AgentRunOutcome::limit_stop` and `::paused`, because those two stop reasons
  carry data and had no constructor — a host reporting either had to write the
  struct literal, which is exactly what makes adding a field source-breaking.
  Migration is now one line.
- `agent_tests_part_02` had grown to 546 lines; the transcript tests move to
  part 03. `caps/agent.rs` went to 527 with the constructors, so the outcome
  types move to `caps/agent/outcome.rs` — the request half and the outcome half
  divide cleanly.

One review point is acknowledged and NOT fixed: an outcome the node turns into
an `Err` — today only `StopReason::Paused` — still loses its transcript, because
the engine builds that error step without a `NodeOutput` to carry one. That is
the run whose transcript is most worth reading. Closing it means giving the
error path somewhere to put one, which is an engine-shaped change; it is named
in the field's docs rather than left to be discovered.
